fix reviews to fulfill specs
